writing-practice/requirements.txt
streamlit==1.32.0
manga-ocr==0.1.10
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
